Refactor app with separate component for the IlearnxApp component
